Almost An Anagram program
An anagram is a word formed by rearranging the letters of another word.
A word is almost an anagram of another word if:
a/ One word is shorter than the other by one letter but otherwise
contains the same letters in any order.
b/ The 2 words are the same length and their character multisets differ
by one character only. e.g.: "aaa" and "aab"

Input:
This program will take 2 words separate by a space in a single line as
input. The world length is anywhere between 1 and 1000 lowercase
characters inclusive.

Output:
1/ str1 is identical to str2
2/ str1 is an anagram of str2
3/ str1 is almost an anagram of str2
4/ str1 is nothing like str2

The first word to be ouput must be the shorter word of the the words are
the same length, or lexicographically least if the length of the 2 words
are the same.

Time limit: 1s